requirements.txt
fastapi==0.115.6
joblib==1.4.2
pandas==2.2.3
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
lightgbm==4.5.0
pytest==8.3.3
uvicorn==0.32.1
